scrapper/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
selenium = "^4.23.1"
nats-py = "^2.8.0"
webdriver-manager = "^4.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
